Extract peer_link_client value types into _client_models.py
Move the 11 module-level types out of
``peer_link_client.py`` (errors, result dataclasses, and
in-session state) into a sibling ``_client_models.py``. The
file used to interleave types with the driver functions that
returned them (``preview_pair`` → ``InitiatorRoundTrip``,
``request_pair`` → ``RequestPairResult``,
``await_pair_status`` → ``PairStatusResult``) plus a cluster
of five errors + one result + two private state dataclasses
sitting between the drivers and the ``PeerLinkClient`` class.
Consolidating them in one wire-shape-only file mirrors the
pattern the rest of the package settled on
(``_validators.py``, ``_summaries.py``, ``_models.py``,
``_storage_codecs.py``).

``peer_link_client.py``: 2101 → 1881 lines (−220).

Types moved:

* Errors: ``PeerLinkClientError``,
  ``PeerLinkNoSessionError``, ``SubmitJobTimeoutError``,
  ``SubmitJobSessionLostError``, ``DownloadArtifactsError``.
* Result dataclasses: ``InitiatorRoundTrip``,
  ``RequestPairResult``, ``PairStatusResult``,
  ``DownloadArtifactsResult``.
* Private session state: ``_DownloadArtifactsState``,
  ``_SessionLoopState``.

External import paths through ``peer_link_client`` keep
working: ``peer_link_client`` imports the names back from
``_client_models`` at the top of the file, so every caller's
``from .peer_link_client import ...`` /
``from esphome_device_builder.controllers.remote_build.peer_link_client import ...``
continues to resolve. No call-site changes elsewhere in the
tree.

Behaviour-preserving. All 3116 tests pass; mypy clean.
